Try to fix Codeclimate issue
